Assignment 3
The goal of this assignment is to learn and practice the concepts covered thus far, in particular: strings (including indexing, slicing, and string methods), control structures (if statements and for-loops), use of range function, function design, and function calls. The only collection I used is strings. No other collections (such as a list/array, tuple, set, or dictionary) were used.