Assignment/Project 5
The goal of this assignment is to learn and practice the concepts covered thus far. In particular, you will get more practice with (2D) lists and functions. For one of the functions, I needed to learn how to prevent syntax errors i.e. crashes by learning about the concept of handling exceptions. NO usage of dictionaries, sets, deque, bisect modules. but, usage of use .sort or sorted functions are implemented.